ROMIO: improve handling of large datatypes
MPI-3 introduced MPI_Count and MPI_Type_size_x.  Use these new routines
throughout ROMIO to support large datatypes.

Still to do:
- update other ADIO drivers

+++ src/mpi/romio/adio/common/ad_coll_exch_new.c
@@ -43,7 +43,7 @@ void ADIOI_Print_flatlist_node(ADIOI_Flatlist_node *flatlist_node_p)
 	return;
     }
     fprintf(stderr, "print flatlist node count = %d (idx,blocklen)\n", 
-	    flatlist_node_p->count);
+	    (int)flatlist_node_p->count);
     for (i = 0; i < flatlist_node_p->count; i++)
     {
 	if (i % 5 == 0 && i != 0)
